+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Addded
+
+- New function: Convert-Metric
+
 ## [1.3.0] - 2023-01-12
 
 ### Fixed